/ -> src
move php files int src folder
add htdocs/index.php
