+++ htdocs/index.php
@@ -1,0 +1,9 @@
+<?php
+
+require_once('../vendor/autoload.php');
+
+use paveld\markdownblog\MarkdownBlog;
+
+$uri = $_SERVER['REQUEST_URI'];
+
+MarkdownBlog::display($uri);
